Refactor into more modular code and add specs
