+++ Gemfile
@@ -3,6 +3,6 @@ source 'https://rubygems.org'
 # Specify your gem's dependencies in dumbo.gemspec
 gemspec
 
-gem 'rspec', '~> 3.0.0'
+gem 'rspec', '~> 3.5.0'
 gem 'pry'
 gem 'rake', '~> 10.5'
