+++ lib/dumbo.rb
@@ -1,22 +1,13 @@
-require 'dumbo/version'
-require 'dumbo/db'
-require 'dumbo/pg_object'
-require 'dumbo/cli'
-require 'dumbo/type'
-require 'dumbo/types/composite_type'
-require 'dumbo/types/enum_type'
-require 'dumbo/types/range_type'
-require 'dumbo/types/base_type'
-require 'dumbo/function'
-require 'dumbo/cast'
-require 'dumbo/aggregate'
-require 'dumbo/dependency_resolver'
-require 'dumbo/extension'
-require 'dumbo/extension_migrator'
-require 'dumbo/extension_version'
-require 'dumbo/operator'
-require 'dumbo/version'
-require 'dumbo/binding_loader'
+require 'thor'
+require 'erubis'
+require 'fileutils'
+require 'pathname'
+
+['', 'command', 'pg_object', 'pg_object/type'].each do |submodule|
+  Dir.glob("#{File.dirname(__FILE__)}/dumbo/#{submodule}/*.rb").each do |path|
+    require File.expand_path(path)
+  end
+end
 
 module Dumbo
   class NoConfigurationError < StandardError
@@ -26,14 +17,45 @@ module Dumbo
   end
 
   class << self
+    def extension_file(*files)
+      File.join(extension_root, *files)
+    end
+
+    def extension_files(*files)
+      Dir.glob(extension_file(*files))
+    end
+
+    def template_root
+      File.join(File.dirname(__FILE__), '..', 'template')
+    end
+
+    def template_files(*files)
+      Dir.glob(File.join(template_root, *files))
+    end
+
     def boot(env)
       raise NoConfigurationError.new(env) if db_config[env].nil?
 
-      DB.connect db_config[env]
+      if !DB.connect(db_config[env])
+        $stderr.puts("Error connecting to PostgreSQL using connection string: `#{connstring(env)}`.")
+        return false
+      end
+
+      true
     end
 
     def db_config
       @config ||= YAML.load_file(File.join('config', 'database.yml'))
+    end
+
+    private
+
+    def connstring(env)
+      db_config[env].map { |key, value| "#{key}=#{value}" }.join(' ')
+    end
+
+    def extension_root
+      FileUtils.pwd
     end
   end
 end
