+++ spec/spec_helper.rb
@@ -1,4 +1,5 @@
 require 'rubygems'
+require 'pry'
 
 $LOAD_PATH.unshift File.expand_path('../..', __FILE__)
 require 'lib/dumbo'
