+++ template/sample.control.erb
@@ -3,3 +3,4 @@ comment = '<%= extension_comment %>'
 default_version = '<%= initial_version %>'
 relocatable = true
 requires = ''
+directory = 'extension'
